CP renderer will be deleted in the manager instead of the panel
